# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for DevContext
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface GitHubUser {
  id: number;
  login: string;
  name?: string | null;
  avatar_url: string;
  html_url: string;
}

export interface Repository {
  id: number;
  name: string;
  full_name: string;
  description?: string | null;
  private: boolean;
  html_url: string;
  updated_at?: string | null;
  language?: string | null;
  stargazers_count: number;
  default_branch: string;
}

export interface Commit {
  sha: string;
  message: string;
  author_name: string;
  author_email: string;
  author_date: string;
  html_url: string;
}

export interface CommitFile {
  filename: string;
  status: string;
  additions: number;
  deletions: number;
  changes: number;
}

export type CommitDetailStats = {
  additions: number;
  deletions: number;
  total: number;
};

export interface CommitDetail {
  sha: string;
  message: string;
  author_name: string;
  author_date: string;
  html_url: string;
  files: CommitFile[];
  stats: CommitDetailStats;
}

export type SummarizeRequestCommitsItem = {
  sha: string;
  message: string;
  author_date: string;
  files: string[];
};

export interface SummarizeRequest {
  repo_name: string;
  commits: SummarizeRequestCommitsItem[];
}

export interface SummarizeResponse {
  what_you_were_doing: string;
  key_changes: string[];
  suggested_next_steps: string[];
  generated_at: string;
}

export type GithubAuthCallbackParams = {
  code: string;
  state?: string;
};

export type ListCommitsParams = {
  per_page?: number;
};