# lib/api-zod/src/generated/types/commit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for DevContext
 * OpenAPI spec version: 0.1.0
 */

export interface Commit {
  sha: string;
  message: string;
  author_name: string;
  author_email: string;
  author_date: string;
  html_url: string;
}