# lib/api-zod/src/generated/types/commitDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for DevContext
 * OpenAPI spec version: 0.1.0
 */
import type { CommitDetailStats } from "./commitDetailStats";
import type { CommitFile } from "./commitFile";

export interface CommitDetail {
  sha: string;
  message: string;
  author_name: string;
  author_date: string;
  html_url: string;
  files: CommitFile[];
  stats: CommitDetailStats;
}