# lib/api-zod/src/generated/types/commitFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for DevContext
 * OpenAPI spec version: 0.1.0
 */

export interface CommitFile {
  filename: string;
  status: string;
  additions: number;
  deletions: number;
  changes: number;
}