# lib/api-zod/src/generated/types/gitHubUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for DevContext
 * OpenAPI spec version: 0.1.0
 */

export interface GitHubUser {
  id: number;
  login: string;
  name?: string | null;
  avatar_url: string;
  html_url: string;
}