# lib/api-zod/src/generated/types/repository.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for DevContext
 * OpenAPI spec version: 0.1.0
 */

export interface Repository {
  id: number;
  name: string;
  full_name: string;
  description?: string | null;
  private: boolean;
  html_url: string;
  updated_at?: string | null;
  language?: string | null;
  stargazers_count: number;
  default_branch: string;
}