# lib/api-zod/src/generated/types/summarizeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for DevContext
 * OpenAPI spec version: 0.1.0
 */
import type { SummarizeRequestCommitsItem } from "./summarizeRequestCommitsItem";

export interface SummarizeRequest {
  repo_name: string;
  commits: SummarizeRequestCommitsItem[];
}