# lib/api-zod/src/generated/types/summarizeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for DevContext
 * OpenAPI spec version: 0.1.0
 */

export interface SummarizeResponse {
  what_you_were_doing: string;
  key_changes: string[];
  suggested_next_steps: string[];
  generated_at: string;
}